refactor: move versioning.py in new utils subpackage
